Fixed issue with Jixel consumer
when musa sends jixel a message, musa must wait for the message to be consumed. For this it uses a map where it stores the event sent to jixel and the associated task (which must complete once jixel receives the message). When the message is consumed, jixel sends an event that completes the task in the flowable process.